moves retrieve_stats_from_import_results to different file
