+++ func_sig_registry/registry/serializers.py
@@ -2,9 +2,11 @@ import json
 
 from rest_framework import serializers
 
+from func_sig_registry.utils.import_statistics import (
+    retrieve_stats_from_import_results,
+)
 from func_sig_registry.utils.abi import (
     is_valid_contract_abi,
-    retrieve_stats_from_import_results,
 )
 from func_sig_registry.utils.events_solidity import (
     normalize_event_signature,
